Move decimal precision and scale formulas to bodo.utils.decimal_utils.
