+++ bodo/libs/decimal_arr_ext.py
@@ -37,6 +37,11 @@ from numba.parfors.array_analysis import ArrayAnalysis
 
 import bodo
 from bodo.libs import decimal_ext
+from bodo.utils.decimal_utils import (
+    decimal_addition_subtraction_output_precision_scale,
+    decimal_division_output_precision_scale,
+    decimal_multiplication_output_precision_scale,
+)
 from bodo.utils.typing import (
     assert_bodo_error,
     get_overload_const_bool,
@@ -1684,19 +1689,6 @@ def overload_sum_decimal_array(arr, parallel=False):
     return impl
 
 
-def decimal_addition_subtraction_output_precision_scale(p1, s1, p2, s2):
-    """
-    Calculate the output precision and scale for a addition/subtraction of two decimals.
-    See: https://docs.snowflake.com/en/sql-reference/operators-arithmetic#addition-and-subtraction
-    """
-    l1 = p1 - s1
-    l2 = p2 - s2
-    l = max(l1, l2) + 1
-    s = max(s1, s2)
-    p = min(l + s, 38)
-    return p, s
-
-
 def add_or_subtract_decimal_scalars(d1, d2, do_addition):  # pragma: no cover
     pass
 
@@ -1926,19 +1918,6 @@ def _add_or_subtract_decimal_arrays(
         ),
         codegen,
     )
-
-
-def decimal_multiplication_output_precision_scale(p1, s1, p2, s2):
-    """
-    Calculate the output precision and scale for a multiplication of two decimals.
-    See: https://docs.snowflake.com/en/sql-reference/operators-arithmetic#multiplication
-    """
-    l1 = p1 - s1
-    l2 = p2 - s2
-    l = l1 + l2
-    s = min(s1 + s2, max(s1, s2, 12))
-    p = min(l + s, 38)
-    return p, s
 
 
 def multiply_decimal_scalars(d1, d2):  # pragma: no cover
@@ -2331,18 +2310,6 @@ def _modulo_decimal_arrays(
         ),
         codegen,
     )
-
-
-def decimal_division_output_precision_scale(p1, s1, p2, s2):
-    """
-    Calculate the output precision and scale for a division of two decimals.
-    See: https://docs.snowflake.com/en/sql-reference/operators-arithmetic#division
-    """
-    l1 = p1 - s1
-    l = l1 + s2
-    s = max(s1, min(s1 + 6, 12))
-    p = min(l + s, 38)
-    return p, s
 
 
 def divide_decimal_scalars(d1, d2, do_div0=False):  # pragma: no cover
